update runCombine.py to fit for smearing of PV
This commit updates the fit to include smearing in the x, y, and z
directions for the primary vertex fit. This is to account for the fact
that we see a big discrepancy between data and MC in the negative
missing mass distribution which is likely caused by the fact that the
primary vertex resolution is worse in data than in MC (this is suggested
by the fact that the p-value of the primary vertex fit does not agree
between data and MC). It also contains lots of other smaller changes:

- make all command line arguments in runCombine.py use the standard form
with lowercase letters and dashes
- fix warnings by calling yaml.load() on newer versions of python
- fix bare cases of raise which was causing an error, and instead always
use raise Exception()
- add a cut on the PV p-value
- remove mu isolation cut
- add cut on sigdxy_vtxD0_PV and pval_D0pismu (both are because these
cuts were relaxed or removed at the ntuplizer level)
- add PV_pval, PV_piK, pval_D0pis to plots
- make sure to close files by always using with open()
- comment out B -> DstDXX (this was a hack to improve the fit and it
doesn't seem to help too much anymore with the primary vertex smearing
added to the fit)
- switch to only running skimmer jobs on condor by default since we are
switching the cluster machines to Alma Linux 8
- fix a bug (I think?) when adding histograms. Previously when adding
the 2D histos there was a line:
-                scale = SM_RDst if 'tau' in n else 1.
+                scale = SM_RDst if 'tau' in key else 1.