fix bug causing B -> D* D events to get the wrong process number
This commit fixes the logic in the skimmer to disambiguate between
different B -> D* D decays. Specifically, there are two decays which
look almost identical except the muon comes from either a D0 or a D+. I
fixed this by adding the muon mother pdg id to the list of conditions to
make sure that we can unambiguously determine the process id.